lite refactoring PiHealthStatistics addRecord() method maintains records collection between min and max records number implemented getRecordsMapSince() and getRecordsSince() methods added method addRecord() to Main class to be executed in a separate thread properties added statistics.records.number.min and statistics.records.number.max

+++ src/main/java/Statistics/PiHealthRecord.java
@@ -4,20 +4,20 @@ package Statistics;
  * Created by dev on 16.05.2017.
  */
 public class PiHealthRecord {
-    private int cpuTemperature;
+    private double cpuTemperature;
     private int cpuFrequency;
     private boolean fanIsOn;
 
     public PiHealthRecord() {
     }
 
-    public PiHealthRecord(int cpuTemperature, int cpuFrequency, boolean fanIsOn) {
+    public PiHealthRecord(double cpuTemperature, int cpuFrequency, boolean fanIsOn) {
         this.cpuTemperature = cpuTemperature;
         this.cpuFrequency = cpuFrequency;
         this.fanIsOn = fanIsOn;
     }
 
-    public int getCpuTemperature() {
+    public double getCpuTemperature() {
         return cpuTemperature;
     }
